fix: VFS daemon crash, TF_KERNEL test gap, add shell command tests
- Fix VFS daemon crash (vfsd.c): zero-initialize msg struct before ipc_recv to prevent dispatch with garbage sender_id/path

- Fix TF_KERNEL test output gap: add build-type stamp to Makefile to prevent stale objects between debug/release builds

- Fix print_report: count test cases (not assertions) in Total

- Fix Logger: add missing '%u' format specifier to vprint

- Add 3 shell command tests: help, echo, uptime (436 total, all pass)

- Add compile_flags.txt for clangd LSP support

+++ os/src/kernel/multiboot2.hpp
@@ -48,9 +48,9 @@ struct MemoryMapTag {
 };
 extern "C" {
     /// @brief Multiboot2 magic value (0x36D76289 if booted by Multiboot2).
-    extern uint32_t multiboot_magic;
+    extern uint64_t multiboot_magic;
     /// @brief Physical pointer to the Multiboot2 info structure.
-    extern uint32_t multiboot_info_ptr;
+    extern uint64_t multiboot_info_ptr;
 }
 
 /// @brief Finds a Multiboot2 tag by type.
